moving modules to modules directoryt
